{"types":"C","runs":59}

+++ Problem15.java
@@ -8,7 +8,7 @@ public class Problem15 extends ConsoleProgram{
 	public void run() {
 		int a = readInt("Enter a: ");
 		int b = readInt("Enter b: ");
-		int average = (a + b) / 2;
+		double average = (a + b) / 2;
 		println("The arithmetic average is "+ average + ".");
 	}
 }
